requirements.txt
Pillow==2.7.0
Unidecode==0.04.16
imgurpython==1.1.3
praw==2.1.21
python-slugify==0.1.0
requests==2.7.0
six==1.9.0
update-checker==0.11
wsgiref==0.1.2
